refactored the code and decoupled the tests
